Update version to 0.9.0-pre  (#2)
* Add base classes to be implemented in ClockWork

* Rename author from Casey Hofland to Cubusky

* Add AnimatorHash

* Add reusable Heatgrid serializers

This includes 2 very useful features:
- GuidAttribute: let's you serialize a string as a valid Guid. `[Guid] public string;`
- OfTypeAttribute: let's you specialize a type in the object selector. Can be used to create an interface selector. `[OfType(typeof(IConstraint))] public Object;`

* Internal guid update

* Add InternalBridge for Engine and Editor

* Update Components folder guid

* Add a TimeSpan drawer

Todo: use the new UXMLElementAttribute API and add TimeOnly drawer once I can upgrade to a Unity version with full .NET integration.

* Allow OfType to specify multiple types

Additionally, Drag and Drop is now respected, the Object Selector does not show up empty and you can drop a game object to select the first available valid component.

* Remove unused SerializedPropertyExtensions

As I cannot vouch for their usability nor accurracy, so it is important to remove this before version 1.0

* Update version to 0.9.0-pre

The only thing left before version 1.0.0 is to improve the documentation page. After that, we feast.