Simplify the Instance implementation
Instance was set up in a weird way that allowed for a component menu to be added easily, but this was cumbersome and unnecessary, as an instance can simply be tested by enabling it. The Internal directory has also be removed, as it only existed to accomodate this simplified instance class, but it would still need to be imported by classes using Instance<T>, defeating the entire purpose of this structure.