Add ReferenceDropdownAttribute
ReferenceDropdownAttribute can be added to a serialized reference to allow you to pick the type you want to use from a dropdown. This makes the SerializeReferenceAttribute incredibly useful, as it allows you to string your code together in a very implementation-agnostic way.

Use like this:
`[SerializeReference, ReferenceDropdown] public IMyInterface myInterface;`