Further Nicify variable names shown in the dropdown
The selected value should show a nicified name, however the dropdown should show the name that will show up when selected as well as the full type name, so that a distinction can be made when choosing between different types.

Additionally, null should not be nicified to distinct it from types called "Null".