Add ISaver
ISaver is an amazing interface that through the ReferenceDropdown will allow users to dynamically select their mode of saving and loading data. It comes implemented with a FileSaverLoader that can save to a file, a DirectorySaverLoader that can save multiple files to a directory, and a TempSaverLoader that keeps the data in a dictionary for as long as the dictionary persists.