Replace UpdateSynchronizer for ManualSynchronizer
Due to the unreliable nature of destructors, `UpdateSynchronizer`'s subscription to `UpdaterService` may not always be removed, leaking code with the ability to cause unintended consequences. You must now process `ManualSynchronizer` queue manually by calling `ManualSynchronizer.ProcessQueue()`.

Remove `UpdaterService`. I never really liked how it would run in the background, and with the changes to the `UpdateSynchronizer` it is no longer needed.